requirements.txt
# ANSYS Fluent Integration
ansys-fluent-core>=0.16.0
pyfluent>=0.16.0

# GitHub Integration
PyGithub>=2.1.1
requests>=2.31.0

# MCP Client
httpx>=0.25.2
websockets>=12.0

# Code Generation & AI
openai>=1.6.0
anthropic>=0.8.0

# CLI Tools
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Utilities
loguru>=0.7.2
tenacity>=8.2.3
pydantic>=2.5.0

# Development & Testing
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
